fix(dsp): inherit caller dtype in VectorizedEnvelopeFollower output (#4225)
process_buffer_vectorized allocated output as float32 regardless of input
dtype, silently downcasting a float64 caller. compressor.py already passes
float32 so there is no behaviour change today, but np.zeros_like(input_levels)
makes the dtype contract correct.

Closes #4225

+++ auralis/dsp/dynamics/vectorized_envelope.py
@@ -66,8 +66,10 @@ class VectorizedEnvelopeFollower:
         if len(input_levels) == 0:
             return np.array([])
 
-        # Allocate output — match input dtype to avoid unnecessary copy
-        output = np.zeros_like(input_levels, dtype=np.float32)
+        # Allocate output inheriting the caller's dtype — forcing float32 here
+        # silently downcast a float64 caller (#4225). compressor.py already
+        # passes float32, so this is contract-correct with no behaviour change.
+        output = np.zeros_like(input_levels)
 
         # Process first sample
         current_env = self.envelope
